# app/static/apiDocs.yaml
swagger: '2.0'
info:
  title: Webhook Delivery System API
  description: API for managing webhook subscriptions and deliveries
  version: '1.0'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /ingest/{id}:
    post:
      tags:
        - webhooks
      summary: Ingest a webhook
      operationId: ingest_webhook
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: The ID of the webhook to ingest
        - name: X-Hub-Signature-256
          in: header
          required: true
          type: string
          description: The signature for the webhook payload
        - in: body
          name: body
          required: true
          schema:
            type: object
            example:
                  event: user.created
                  data:
                    id: 123
                    name: John Doe
                  timestamp: 2025-04-27T10:30:00Z
      responses:
        200:
          description: Webhook ingested successfully
          schema:
            $ref: '#/definitions/Webhook'
        400:
          description: Invalid input
  /ingest/bypass-signature/{id}:
    post:
      tags:
        - webhooks
      summary: Ingest a webhook without signature verification
      operationId: ingest_webhook_bypass_signature
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: The ID of the webhook to ingest
        - in: body
          name: body
          required: true
          schema:
            type: object
            example:
                  event: user.created
                  data:
                    id: 123
                    name: John Doe
                  timestamp: 2025-04-27T10:30:00Z
      responses:
        200:
          description: Webhook ingested successfully without signature verification
          schema:
            $ref: '#/definitions/Webhook'
        400:
          description: Invalid input
...
definitions:
...
  Webhook:
    type: object
    properties:
      id:
        type: string
        description: Webhook ID (UUID)
        example: 123e4567-e89b-12d3-a456-426614174000
      subscription_id:
        type: integer
        description: Subscription ID
        example: 1
      payload:
        type: object
        description: Webhook payload (JSON)
        example:
          event: user.created
          data:
            id: 123
            name: John Doe
          timestamp: 2025-04-27T10:30:00Z
      received_at:
        type: string
        format: date-time
        description: Reception timestamp
        example: 2025-04-27T10:30:00Z
      signature_verified:
        type: boolean
        description: Whether signature was verified
        example: true
...